# perf(pycodestyle): Reduce allocations when computing logical lines