--- a/crates/ruff/src/checkers/logical_lines.rs
+++ b/crates/ruff/src/checkers/logical_lines.rs
@@ -1,6 +1,5 @@
 #![allow(dead_code, unused_imports, unused_variables)]
 
-use bisection::bisect_left;
 use itertools::Itertools;
 use rustpython_parser::ast::Location;
 use rustpython_parser::lexer::LexResult;
@@ -10,7 +9,7 @@ use ruff_python_ast::source_code::{Locator, Stylist};
 use ruff_python_ast::types::Range;
 
 use crate::registry::{AsRule, Rule};
-use crate::rules::pycodestyle::logical_lines::{iter_logical_lines, TokenFlags};
+use crate::rules::pycodestyle::logical_lines::{LogicalLines, TokenFlags};
 use crate::rules::pycodestyle::rules::{
     extraneous_whitespace, indentation, missing_whitespace, missing_whitespace_after_keyword,
     missing_whitespace_around_operator, space_around_operator, whitespace_around_keywords,
@@ -20,23 +19,18 @@ use crate::rules::pycodestyle::rules::{
 use crate::settings::{flags, Settings};
 
 /// Return the amount of indentation, expanding tabs to the next multiple of 8.
-fn expand_indent(mut line: &str) -> usize {
-    while line.ends_with("\n\r") {
-        line = &line[..line.len() - 2];
-    }
-    if !line.contains('\t') {
-        return line.len() - line.trim_start().len();
-    }
+fn expand_indent(line: &str) -> usize {
+    let line = line.trim_end_matches(['\n', '\r']);
+
     let mut indent = 0;
-    for c in line.chars() {
-        if c == '\t' {
-            indent = (indent / 8) * 8 + 8;
-        } else if c == ' ' {
-            indent += 1;
-        } else {
-            break;
+    for c in line.bytes() {
+        match c {
+            b'\t' => indent = (indent / 8) * 8 + 8,
+            b' ' => indent += 1,
+            _ => break,
         }
     }
+
     indent
 }
 
@@ -52,25 +46,18 @@ pub fn check_logical_lines(
     let indent_char = stylist.indentation().as_char();
     let mut prev_line = None;
     let mut prev_indent_level = None;
-    for line in iter_logical_lines(tokens, locator) {
-        if line.mapping.is_empty() {
-            continue;
-        }
-
+    for line in &LogicalLines::from_tokens(tokens, locator) {
         // Extract the indentation level.
-        let start_loc = line.mapping[0].1;
-        let start_line = locator.slice(Range::new(Location::new(start_loc.row(), 0), start_loc));
+        let Some(start_loc) = line.first_token_location() else { continue; };
+        let start_line = locator.slice(Range::new(Location::new(start_loc.row(), 0), *start_loc));
         let indent_level = expand_indent(start_line);
         let indent_size = 4;
 
-        // Generate mapping from logical to physical offsets.
-        let mapping_offsets = line.mapping.iter().map(|(offset, _)| *offset).collect_vec();
-
-        if line.flags.contains(TokenFlags::OPERATOR) {
-            for (index, kind) in space_around_operator(&line.text) {
-                let (token_offset, pos) = line.mapping[bisect_left(&mapping_offsets, &index)];
-                let location = Location::new(pos.row(), pos.column() + index - token_offset);
+        if line.flags().contains(TokenFlags::OPERATOR) {
+            for (index, kind) in space_around_operator(line.text()) {
                 if settings.rules.enabled(kind.rule()) {
+                    let (token_offset, pos) = line.mapping(index);
+                    let location = Location::new(pos.row(), pos.column() + index - token_offset);
                     diagnostics.push(Diagnostic {
                         kind,
                         location,
@@ -82,13 +69,13 @@ pub fn check_logical_lines(
             }
         }
         if line
-            .flags
+            .flags()
             .contains(TokenFlags::OPERATOR | TokenFlags::PUNCTUATION)
         {
-            for (index, kind) in extraneous_whitespace(&line.text) {
-                let (token_offset, pos) = line.mapping[bisect_left(&mapping_offsets, &index)];
-                let location = Location::new(pos.row(), pos.column() + index - token_offset);
+            for (index, kind) in extraneous_whitespace(line.text()) {
                 if settings.rules.enabled(kind.rule()) {
+                    let (token_offset, pos) = line.mapping(index);
+                    let location = Location::new(pos.row(), pos.column() + index - token_offset);
                     diagnostics.push(Diagnostic {
                         kind,
                         location,
@@ -99,11 +86,11 @@ pub fn check_logical_lines(
                 }
             }
         }
-        if line.flags.contains(TokenFlags::KEYWORD) {
-            for (index, kind) in whitespace_around_keywords(&line.text) {
-                let (token_offset, pos) = line.mapping[bisect_left(&mapping_offsets, &index)];
-                let location = Location::new(pos.row(), pos.column() + index - token_offset);
+        if line.flags().contains(TokenFlags::KEYWORD) {
+            for (index, kind) in whitespace_around_keywords(line.text()) {
                 if settings.rules.enabled(kind.rule()) {
+                    let (token_offset, pos) = line.mapping(index);
+                    let location = Location::new(pos.row(), pos.column() + index - token_offset);
                     diagnostics.push(Diagnostic {
                         kind,
                         location,
@@ -114,7 +101,7 @@ pub fn check_logical_lines(
                 }
             }
 
-            for (location, kind) in missing_whitespace_after_keyword(&line.tokens) {
+            for (location, kind) in missing_whitespace_after_keyword(line.tokens()) {
                 if settings.rules.enabled(kind.rule()) {
                     diagnostics.push(Diagnostic {
                         kind,
@@ -126,8 +113,8 @@ pub fn check_logical_lines(
                 }
             }
         }
-        if line.flags.contains(TokenFlags::COMMENT) {
-            for (range, kind) in whitespace_before_comment(&line.tokens, locator) {
+        if line.flags().contains(TokenFlags::COMMENT) {
+            for (range, kind) in whitespace_before_comment(line.tokens(), locator) {
                 if settings.rules.enabled(kind.rule()) {
                     diagnostics.push(Diagnostic {
                         kind,
@@ -139,9 +126,9 @@ pub fn check_logical_lines(
                 }
             }
         }
-        if line.flags.contains(TokenFlags::OPERATOR) {
+        if line.flags().contains(TokenFlags::OPERATOR) {
             for (location, kind) in
-                whitespace_around_named_parameter_equals(&line.tokens, &line.text)
+                whitespace_around_named_parameter_equals(line.tokens(), line.text())
             {
                 if settings.rules.enabled(kind.rule()) {
                     diagnostics.push(Diagnostic {
@@ -153,7 +140,7 @@ pub fn check_logical_lines(
                     });
                 }
             }
-            for (location, kind) in missing_whitespace_around_operator(&line.tokens) {
+            for (location, kind) in missing_whitespace_around_operator(line.tokens()) {
                 if settings.rules.enabled(kind.rule()) {
                     diagnostics.push(Diagnostic {
                         kind,
@@ -172,7 +159,7 @@ pub fn check_logical_lines(
             let should_fix = false;
 
             for diagnostic in
-                missing_whitespace(&line.text, start_loc.row(), should_fix, indent_level)
+                missing_whitespace(line.text(), start_loc.row(), should_fix, indent_level)
             {
                 if settings.rules.enabled(diagnostic.kind.rule()) {
                     diagnostics.push(diagnostic);
@@ -180,7 +167,7 @@ pub fn check_logical_lines(
             }
         }
 
-        if line.flags.contains(TokenFlags::BRACKET) {
+        if line.flags().contains(TokenFlags::BRACKET) {
             #[cfg(feature = "logical_lines")]
             let should_fix =
                 autofix.into() && settings.rules.should_fix(Rule::WhitespaceBeforeParameters);
@@ -188,7 +175,7 @@ pub fn check_logical_lines(
             #[cfg(not(feature = "logical_lines"))]
             let should_fix = false;
 
-            for diagnostic in whitespace_before_parameters(&line.tokens, should_fix) {
+            for diagnostic in whitespace_before_parameters(line.tokens(), should_fix) {
                 if settings.rules.enabled(diagnostic.kind.rule()) {
                     diagnostics.push(diagnostic);
                 }
@@ -203,7 +190,7 @@ pub fn check_logical_lines(
             prev_indent_level,
             indent_size,
         ) {
-            let (token_offset, pos) = line.mapping[bisect_left(&mapping_offsets, &index)];
+            let (token_offset, pos) = line.mapping(index);
             let location = Location::new(pos.row(), pos.column() + index - token_offset);
             if settings.rules.enabled(kind.rule()) {
                 diagnostics.push(Diagnostic {
@@ -229,9 +216,8 @@ mod tests {
     use rustpython_parser::lexer::LexResult;
     use rustpython_parser::{lexer, Mode};
 
+    use crate::rules::pycodestyle::logical_lines::LogicalLines;
     use ruff_python_ast::source_code::Locator;
-
-    use crate::checkers::logical_lines::iter_logical_lines;
 
     #[test]
     fn split_logical_lines() {
@@ -241,9 +227,9 @@ y = 2
 z = x + 1"#;
         let lxr: Vec<LexResult> = lexer::lex(contents, Mode::Module).collect();
         let locator = Locator::new(contents);
-        let actual: Vec<String> = iter_logical_lines(&lxr, &locator)
+        let actual: Vec<String> = LogicalLines::from_tokens(&lxr, &locator)
             .into_iter()
-            .map(|line| line.text)
+            .map(|line| line.text().to_string())
             .collect();
         let expected = vec![
             "x = 1".to_string(),
@@ -262,9 +248,9 @@ y = 2
 z = x + 1"#;
         let lxr: Vec<LexResult> = lexer::lex(contents, Mode::Module).collect();
         let locator = Locator::new(contents);
-        let actual: Vec<String> = iter_logical_lines(&lxr, &locator)
+        let actual: Vec<String> = LogicalLines::from_tokens(&lxr, &locator)
             .into_iter()
-            .map(|line| line.text)
+            .map(|line| line.text().to_string())
             .collect();
         let expected = vec![
             "x = [1, 2, 3, ]".to_string(),
@@ -276,9 +262,9 @@ z = x + 1"#;
         let contents = "x = 'abc'";
         let lxr: Vec<LexResult> = lexer::lex(contents, Mode::Module).collect();
         let locator = Locator::new(contents);
-        let actual: Vec<String> = iter_logical_lines(&lxr, &locator)
+        let actual: Vec<String> = LogicalLines::from_tokens(&lxr, &locator)
             .into_iter()
-            .map(|line| line.text)
+            .map(|line| line.text().to_string())
             .collect();
         let expected = vec!["x = \"xxx\"".to_string()];
         assert_eq!(actual, expected);
@@ -289,9 +275,9 @@ def f():
 f()"#;
         let lxr: Vec<LexResult> = lexer::lex(contents, Mode::Module).collect();
         let locator = Locator::new(contents);
-        let actual: Vec<String> = iter_logical_lines(&lxr, &locator)
+        let actual: Vec<String> = LogicalLines::from_tokens(&lxr, &locator)
             .into_iter()
-            .map(|line| line.text)
+            .map(|line| line.text().to_string())
             .collect();
         let expected = vec!["def f():", "x = 1", "f()"];
         assert_eq!(actual, expected);
@@ -304,9 +290,9 @@ def f():
 f()"#;
         let lxr: Vec<LexResult> = lexer::lex(contents, Mode::Module).collect();
         let locator = Locator::new(contents);
-        let actual: Vec<String> = iter_logical_lines(&lxr, &locator)
+        let actual: Vec<String> = LogicalLines::from_tokens(&lxr, &locator)
             .into_iter()
-            .map(|line| line.text)
+            .map(|line| line.text().to_string())
             .collect();
         let expected = vec!["def f():", "\"xxxxxxxxxxxxxxxxxxxx\"", "", "x = 1", "f()"];
         assert_eq!(actual, expected);