--- a/crates/ruff/src/rules/pycodestyle/logical_lines.rs
+++ b/crates/ruff/src/rules/pycodestyle/logical_lines.rs
@@ -2,6 +2,8 @@ use bitflags::bitflags;
 use rustpython_parser::ast::Location;
 use rustpython_parser::lexer::LexResult;
 use rustpython_parser::Tok;
+use std::borrow::Cow;
+use std::iter::Flatten;
 use unicode_width::UnicodeWidthStr;
 
 use ruff_python_ast::source_code::Locator;
@@ -11,7 +13,7 @@ use crate::rules::pycodestyle::helpers::{is_keyword_token, is_op_token};
 
 bitflags! {
     #[derive(Default)]
-    pub struct TokenFlags: u32 {
+    pub struct TokenFlags: u8 {
         /// Whether the logical line contains an operator.
         const OPERATOR = 0b0000_0001;
         /// Whether the logical line contains a bracket.
@@ -39,64 +41,75 @@ impl<'a> LogicalLine<'a> {
     }
 }
 
-fn build_line<'a>(
+struct LineBuilder<'a> {
     tokens: Vec<(Location, &'a Tok, Location)>,
-    locator: &Locator,
-) -> LogicalLine<'a> {
-    let mut logical = String::with_capacity(88);
-    let mut mapping = Vec::new();
-    let mut flags = TokenFlags::empty();
-    let mut prev: Option<&Location> = None;
-    let mut length = 0;
-    for (start, tok, end) in &tokens {
+    // BTreeMap?
+    mappings: Vec<(usize, Location)>,
+    text: String,
+    flags: TokenFlags,
+    previous: Option<Location>,
+}
+
+impl<'a> LineBuilder<'a> {
+    fn new() -> Self {
+        Self {
+            tokens: Vec::with_capacity(32),
+            mappings: Vec::new(),
+            text: String::with_capacity(88),
+            flags: TokenFlags::empty(),
+            previous: None,
+        }
+    }
+
+    fn push_token(&mut self, start: Location, token: &'a Tok, end: Location, locator: &Locator) {
+        self.tokens.push((start, token, end));
+
         if matches!(
-            tok,
+            token,
             Tok::Newline | Tok::NonLogicalNewline | Tok::Indent | Tok::Dedent
         ) {
-            continue;
+            return;
         }
 
-        if mapping.is_empty() {
-            mapping.push((0, *start));
+        if self.mappings.is_empty() {
+            self.mappings.push((0, start));
         }
 
-        if matches!(tok, Tok::Comment { .. }) {
-            flags.insert(TokenFlags::COMMENT);
-            continue;
+        if matches!(token, Tok::Comment(..)) {
+            self.flags.insert(TokenFlags::COMMENT);
+            return;
         }
 
-        if is_op_token(tok) {
-            flags.insert(TokenFlags::OPERATOR);
+        if is_op_token(token) {
+            self.flags.insert(TokenFlags::OPERATOR);
         }
 
         if matches!(
-            tok,
+            token,
             Tok::Lpar | Tok::Lsqb | Tok::Lbrace | Tok::Rpar | Tok::Rsqb | Tok::Rbrace
         ) {
-            flags.insert(TokenFlags::BRACKET);
+            self.flags.insert(TokenFlags::BRACKET);
         }
 
-        if matches!(tok, Tok::Comma | Tok::Semi | Tok::Colon) {
-            flags.insert(TokenFlags::PUNCTUATION);
+        if matches!(token, Tok::Comma | Tok::Semi | Tok::Colon) {
+            self.flags.insert(TokenFlags::PUNCTUATION);
         }
 
-        if is_keyword_token(tok) {
-            flags.insert(TokenFlags::KEYWORD);
+        if is_keyword_token(token) {
+            self.flags.insert(TokenFlags::KEYWORD);
         }
 
         // TODO(charlie): "Mute" strings.
-        let s;
-        let text = if let Tok::String { value, .. } = tok {
-            s = format!("\"{}\"", "x".repeat(value.width()).clone());
-            &s
+        let text = if let Tok::String { value, .. } = token {
+            Cow::Owned(format!("\"{}\"", "x".repeat(value.width())))
         } else {
-            locator.slice(Range {
-                location: *start,
-                end_location: *end,
-            })
+            Cow::Borrowed(locator.slice(Range {
+                location: start,
+                end_location: end,
+            }))
         };
 
-        if let Some(prev) = prev {
+        if let Some(prev) = self.previous {
             if prev.row() != start.row() {
                 let prev_text = locator.slice(Range {
                     location: Location::new(prev.row(), prev.column() - 1),
@@ -106,61 +119,96 @@ fn build_line<'a>(
                     || ((prev_text != "{" && prev_text != "[" && prev_text != "(")
                         && (text != "}" && text != "]" && text != ")"))
                 {
-                    logical.push(' ');
-                    length += 1;
+                    self.text.push(' ');
                 }
             } else if prev.column() != start.column() {
                 let prev_text = locator.slice(Range {
-                    location: *prev,
-                    end_location: *start,
+                    location: prev,
+                    end_location: start,
                 });
-                logical.push_str(prev_text);
-                length += prev_text.len();
+                self.text.push_str(prev_text);
             }
         }
-        logical.push_str(text);
-        length += text.len();
-        mapping.push((length, *end));
-        prev = Some(end);
+
+        self.text.push_str(&text);
+        self.mappings.push((self.text.len(), end));
+        self.previous = Some(end);
     }
 
-    LogicalLine {
-        text: logical,
-        mapping,
-        flags,
-        tokens,
+    fn is_empty(&self) -> bool {
+        self.tokens.is_empty()
+    }
+
+    fn finish(self) -> LogicalLine<'a> {
+        LogicalLine {
+            text: self.text,
+            mapping: self.mappings,
+            flags: self.flags,
+            tokens: self.tokens,
+        }
     }
 }
 
-pub fn iter_logical_lines<'a>(tokens: &'a [LexResult], locator: &Locator) -> Vec<LogicalLine<'a>> {
-    let mut parens = 0;
-    let mut accumulator = Vec::with_capacity(32);
-    let mut lines = Vec::with_capacity(128);
-    for &(start, ref tok, end) in tokens.iter().flatten() {
-        accumulator.push((start, tok, end));
-        if matches!(tok, Tok::Lbrace | Tok::Lpar | Tok::Lsqb) {
-            parens += 1;
-        } else if matches!(tok, Tok::Rbrace | Tok::Rpar | Tok::Rsqb) {
-            parens -= 1;
-        } else if parens == 0 {
-            if matches!(
-                tok,
-                Tok::Newline | Tok::NonLogicalNewline | Tok::Comment(..)
-            ) {
-                if matches!(tok, Tok::Newline) {
-                    lines.push(build_line(accumulator, locator));
-                    accumulator = Vec::with_capacity(32);
-                } else if tokens.len() == 1 {
-                    accumulator.remove(0);
-                } else {
-                    lines.push(build_line(accumulator, locator));
-                    accumulator = Vec::with_capacity(32);
-                }
-            }
+pub fn iter_logical_lines<'a>(
+    tokens: &'a [LexResult],
+    locator: &'a Locator,
+) -> LogicalLinesIterator<'a> {
+    LogicalLinesIterator::new(tokens, locator)
+}
+
+pub struct LogicalLinesIterator<'a> {
+    tokens: Flatten<std::slice::Iter<'a, LexResult>>,
+    locator: &'a Locator<'a>,
+    parens: usize,
+    single_item: bool,
+}
+
+impl<'a> LogicalLinesIterator<'a> {
+    fn new(tokens: &'a [LexResult], locator: &'a Locator<'a>) -> Self {
+        Self {
+            single_item: tokens.len() == 1,
+            locator,
+            parens: 0,
+            tokens: tokens.iter().flatten(),
         }
     }
-    if !accumulator.is_empty() {
-        lines.push(build_line(accumulator, locator));
+}
+
+impl<'a> Iterator for LogicalLinesIterator<'a> {
+    type Item = LogicalLine<'a>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let mut line_builder = LineBuilder::new();
+
+        for (start, token, end) in self.tokens.by_ref() {
+            line_builder.push_token(*start, token, *end, &self.locator);
+
+            match token {
+                Tok::Lbrace | Tok::Lpar | Tok::Lsqb => {
+                    self.parens += 1;
+                }
+                Tok::Rbrace | Tok::Rpar | Tok::Rsqb => {
+                    self.parens -= 1;
+                }
+                Tok::Newline | Tok::NonLogicalNewline | Tok::Comment(_) if self.parens == 0 => {
+                    return if matches!(token, Tok::Newline) {
+                        Some(line_builder.finish())
+                    }
+                    // Comment only file or non logical new line?
+                    else if self.single_item {
+                        None
+                    } else {
+                        Some(line_builder.finish())
+                    };
+                }
+                _ => {}
+            }
+        }
+
+        if !line_builder.is_empty() {
+            Some(line_builder.finish())
+        } else {
+            None
+        }
     }
-    lines
 }