This PR refactors the logical lines computation that the pycodestyle rules use from using `O(lines)` allocations to `O(1)`.

The current implementation allocates for every line:
* a `String` for the line's text
* a `Vec` for the logical-line text `offset` to `Location` mappings
* a `Vec` for the line's token

This PR reduces the allocations by allocating a single `String` for the text, two Vec's for the mappings and tokens, and one `Vec` for storing the line-metadata. The implementation slices into the shared text, mappings, and tokens structures to get the *view* for a single line. 

Allocating the data structures for all lines has the advantage of reserving the containers with the right capacity because we know the number of tokens. This avoids the *guessing* of the old implementation that used some more-or-less arbitrary numbers for the tokens-capacity. 

## Performance

* `no-logical`: Logical lines feature disabled
* `base-logical`: Logical lines feature enabled on main
* `pr3715`: This branch

```
group                                      base-logical                           no-logical                             pr3715
-----                                      ------------                           ----------                             ------
linter/all-rules/large/dataset.py          1.15      9.8±0.02ms     4.2 MB/sec    1.00      8.5±0.01ms     4.8 MB/sec    1.10      9.4±0.15ms     4.3 MB/sec
linter/all-rules/numpy/ctypeslib.py        1.14      2.4±0.04ms     6.8 MB/sec    1.00      2.1±0.01ms     7.8 MB/sec    1.15      2.5±0.00ms     6.8 MB/sec
linter/all-rules/numpy/globals.py          1.21    301.1±1.97µs     9.8 MB/sec    1.00    247.9±3.25µs    11.9 MB/sec    1.16    286.9±3.46µs    10.3 MB/sec
linter/all-rules/pydantic/types.py         1.19      4.4±0.03ms     5.8 MB/sec    1.00      3.7±0.04ms     7.0 MB/sec    1.19      4.4±0.02ms     5.9 MB/sec
linter/default-rules/large/dataset.py      1.28      6.0±0.08ms     6.8 MB/sec    1.00      4.7±0.01ms     8.7 MB/sec    1.24      5.8±0.08ms     7.0 MB/sec
linter/default-rules/numpy/ctypeslib.py    1.32   1317.7±3.83µs    12.6 MB/sec    1.00    999.5±5.57µs    16.7 MB/sec    1.31  1304.4±16.73µs    12.8 MB/sec
linter/default-rules/numpy/globals.py      1.44    145.7±1.14µs    20.3 MB/sec    1.00    101.0±0.41µs    29.2 MB/sec    1.44    145.7±1.31µs    20.2 MB/sec
linter/default-rules/pydantic/types.py     1.32      2.8±0.04ms     9.0 MB/sec    1.00      2.1±0.01ms    11.9 MB/sec    1.26      2.7±0.03ms     9.5 MB/sec
```

This change improves performance overall but the pycodestyle still add a significant overhead.